move HasRoles to trait folder, update person controller test to use delete-duplicate permission
