--- app/Http/Requests/UpdateUomRequest0.php
@@ -1,46 +1,0 @@
-<?php
-
-namespace App\Http\Requests;
-
-use Illuminate\Foundation\Http\FormRequest;
-use Illuminate\Validation\Rule;
-
-class UpdateUomRequest0 extends FormRequest
-{
-    /**
-     * Determine if the user is authorized to make this request.
-     *
-     * @return bool
-     */
-    public function authorize()
-    {
-        return true;
-    }
-
-    /**
-     * Get the validation rules that apply to the request.
-     *
-     * @return array
-     */
-    public function rules()
-    {
-        return [
-            'id' => 'integer|min:1|required',
-            'type' => 'string|required',
-            'unit_name' => 'string|max:125|nullable',
-            'unit_symbol' => 'string|max:125|nullable',
-            'description' => 'string|nullable',
-            'is_active' => 'boolean|nullable',
-        ];
-    }
-
-    /**
-     * Get the error messages for the defined validation rules.
-     *
-     * @return array
-     */
-    public function messages()
-    {
-        return [];
-    }
-}
