+++ app/UserRepository.php
@@ -1,6 +1,7 @@
 <?php
 
 namespace App;
+use App\Models\User;
 
 class UserRepository
 {
